Move Asset Tests into it's own file
